+++ changelog.md
@@ -26,8 +26,9 @@ All notable changes to this project will be documented in this file.
 - Override the `toString` function of `NotBlankString`
   ([#117](https://github.com/kotools/types/issues/117)) and `NotEmptySet`
   ([#123](https://github.com/kotools/types/issues/123)).
-- Convert the `NotBlankString` type from a class to a sealed interface
-  ([#124](https://github.com/kotools/types/issues/124)).
+- Convert the following types from a class to a sealed interface:
+    - `NotBlankString` ([#124](https://github.com/kotools/types/issues/124))
+    - `NotEmptyList` ([#125](https://github.com/kotools/types/issues/125)).
 
 ## Version 3.0.0 - 2022/10/16